Changed lo_discount predicate from int64_t to uint8_t

+++ src/ssb/ssb_workload.cpp
@@ -76,7 +76,7 @@ SSB::SSB(int SF, bool print) {
 
     reset_results();
 
-    auto field1 = arrow::field("order key", arrow::int64());
+    auto field1 = arrow::field("order key", arrow::uint32());
     auto field2 = arrow::field("line number", arrow::int64());
     auto field3 = arrow::field("cust key", arrow::int64());
     auto field4 = arrow::field("part key", arrow::int64());
@@ -87,7 +87,7 @@ SSB::SSB(int SF, bool print) {
     auto field9 = arrow::field("quantity", arrow::int64());
     auto field10 = arrow::field("extended price", arrow::int64());
     auto field11 = arrow::field("ord total price", arrow::int64());
-    auto field12 = arrow::field("discount", arrow::int64());
+    auto field12 = arrow::field("discount", arrow::uint8());
     auto field13 = arrow::field("revenue", arrow::int64());
     auto field14 = arrow::field("supply cost", arrow::int64());
     auto field15 = arrow::field("tax", arrow::int64());
@@ -251,7 +251,7 @@ void SSB::q11() {
         {lo,
          "discount"},
         arrow::compute::CompareOperator::GREATER_EQUAL,
-        arrow::Datum((int64_t) 1)
+        arrow::Datum((uint8_t) 1)
     };
     auto discount_pred_node_1 =
         std::make_shared<PredicateNode>(
@@ -262,7 +262,7 @@ void SSB::q11() {
         {lo,
          "discount"},
         arrow::compute::CompareOperator::LESS_EQUAL,
-        arrow::Datum((int64_t) 3)
+        arrow::Datum((uint8_t) 3)
     };
     auto discount_pred_node_2 =
         std::make_shared<PredicateNode>(
@@ -362,7 +362,7 @@ void SSB::q12() {
         {lo,
          "discount"},
         arrow::compute::CompareOperator::GREATER_EQUAL,
-        arrow::Datum((int64_t) 4)
+        arrow::Datum((uint8_t) 4)
     };
     auto discount_pred_node_1 =
         std::make_shared<PredicateNode>(
@@ -373,7 +373,7 @@ void SSB::q12() {
         {lo,
          "discount"},
         arrow::compute::CompareOperator::LESS_EQUAL,
-        arrow::Datum((int64_t) 6)
+        arrow::Datum((uint8_t) 6)
     };
     auto discount_pred_node_2 =
         std::make_shared<PredicateNode>(
@@ -493,7 +493,7 @@ void SSB::q13() {
         {lo,
          "discount"},
         arrow::compute::CompareOperator::GREATER_EQUAL,
-        arrow::Datum((int64_t) 5)
+        arrow::Datum((uint8_t) 5)
     };
     auto discount_pred_node_1 =
         std::make_shared<PredicateNode>(
@@ -504,7 +504,7 @@ void SSB::q13() {
         {lo,
          "discount"},
         arrow::compute::CompareOperator::LESS_EQUAL,
-        arrow::Datum((int64_t) 7)
+        arrow::Datum((uint8_t) 7)
     };
     auto discount_pred_node_2 =
         std::make_shared<PredicateNode>(
